requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
astroid==3.3.8
asyncpg==0.30.0
bcrypt==4.0.1
certifi==2025.1.31
click==8.1.8
dill==0.3.9
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.8
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
isort==6.0.0
Mako==1.3.9
MarkupSafe==3.0.2
mccabe==0.7.0
packaging==24.2
passlib==1.7.4
platformdirs==4.3.6
pluggy==1.5.0
psycopg==3.2.4
psycopg-binary==3.2.4
pyasn1==0.6.1
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydentic==0.0.1.dev3
PyJWT==2.10.1
pylint==3.3.4
pytest==8.3.4
pytest-asyncio==0.25.3
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
python-stdnum==1.20
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
tomlkit==0.13.2
typing_extensions==4.12.2
uvicorn==0.34.0
